+++ internal/css_parser/css_parser.go
@@ -1144,12 +1144,13 @@ abortRuleParser:
 		if p.peek(css_lexer.TIdent) {
 			name = p.decoded()
 			p.advance()
-		} else if p.eat(css_lexer.TString) {
-			// Consider string names to be an unknown rule even though they are allowed
-			// by the specification and they work in Firefox because they do not work in
-			// Chrome or Safari. We don't take the effort to support this Firefox-only
-			// feature natively. Instead, we just pass the syntax through unmodified.
-			break
+		} else if p.peek(css_lexer.TString) {
+			// Note: Strings as names is allowed in the CSS specification and works in
+			// Firefox and Safari but Chrome has strangely decided to deliberately not
+			// support this. We always turn all string names into identifiers to avoid
+			// them silently breaking in Chrome.
+			name = p.decoded()
+			p.advance()
 		} else if !p.expect(css_lexer.TIdent) {
 			break
 		}